refactor(parser): split streaming source preprocessing
