+++ packages/markdown-parser/src/parser/index.ts
@@ -1,27 +1,15 @@
 import type { MarkdownIt, Token } from '../markdown-it-types'
 import type { HtmlBlockNode, InternalParseOptions, MarkdownToken, ParsedNode, ParseOptions } from '../types'
 import { normalizeCustomHtmlTags } from '../customHtmlTags'
-import { NON_STRUCTURING_HTML_TAGS, STANDARD_BLOCK_HTML_TAGS, VOID_HTML_TAGS } from '../htmlTags'
+import { NON_STRUCTURING_HTML_TAGS, VOID_HTML_TAGS } from '../htmlTags'
 import { escapeTagForRegExp, findTagCloseIndexOutsideQuotes, parseTagAttrs } from '../htmlTagUtils'
-import { isMathLike } from '../plugins/isMathLike'
 import { isCacheStableLinkValidator, readSyntheticLinkOrigin } from '../plugins/linkTokenMetadata'
 import {
-  getTolerantMathBlockBoundaryStreamKey,
   hasMarkstreamMathPlugin,
   mayContainTolerantMathBlockBoundaryOpener,
 } from '../plugins/math'
 import { parseInlineTokens } from './inline-parsers'
 import { createLinkifyDemotionContextTracker } from './linkifyHeuristics'
-import {
-  consumeMarkdownIndent,
-  countRepeatedChar,
-  findCodeSpanCloseIndex,
-  getMarkdownIndent,
-  isEscapedDelimiterAt,
-  matchMarkdownFenceMarker,
-  stripMarkdownBlockquotePrefix,
-  stripMarkdownListPrefix,
-} from './markdown-context'
 import { parseCommonBlockToken } from './node-parsers/block-token-parser'
 import { parseBlockquote } from './node-parsers/blockquote-parser'
 import { containerTokenHandlers } from './node-parsers/container-token-handlers'
@@ -31,6 +19,12 @@ import { parseList } from './node-parsers/list-parser'
 import { parseParagraph } from './node-parsers/paragraph-parser'
 import { applyNodeSourceMap, createSourceMapFromOffsets } from './node-source-map'
 import { createSourceLineMapper } from './source-line-mapper'
+import {
+  clearTolerantMathBoundaryStreamCache,
+  shouldUseSyncParseForPendingTolerantMathBoundary,
+  syncTolerantMathBoundaryStreamCache,
+} from './streaming/boundary-state'
+import { clearSafeMarkdownCache, getSafeMarkdown } from './streaming/safe-markdown'
 import { cloneMarkdownTokens } from './token-clone'
 
 type ParsedNodeWithFields = ParsedNode & {
@@ -78,62 +72,6 @@ function recordInternalNodeSourceRange(node: ParsedNode, token: MarkdownToken | 
     end: offsets[Math.max(0, Math.trunc(endLine))] ?? source.length,
   })
 }
-
-interface ExplicitBracketMathContext {
-  fenceChar: '`' | '~' | ''
-  fenceInBlockquote: boolean
-  fenceInList: boolean
-  fenceLen: number
-  fenceListIndent: number
-  inDollarMath: boolean
-  inFence: boolean
-  inMath: boolean
-  listContentIndent: number | null
-  dollarMathOpenOffset: number | null
-  mathOpenOffset: number | null
-}
-
-interface ExplicitBracketMathStreamState {
-  committedContext: ExplicitBracketMathContext
-  context: ExplicitBracketMathContext
-  lineBuffer: string
-}
-
-const tolerantMathBoundaryStreamCache = new WeakMap<object, {
-  explicitBracketMath: ExplicitBracketMathStreamState
-  source: string
-  key: string | null
-  pendingCandidate: boolean
-}>()
-const pendingExplicitMathTailCache = new WeakMap<object, {
-  source: string
-  state: ExplicitBracketMathStreamState
-}>()
-
-const TOLERANT_BOUNDARY_SPLIT_OPENERS = ['$', '\\[']
-const STREAMING_ADMONITION_OPEN_RE = /(^|\r?\n)[\t ]*:::[\t ]*(?:warning|info|note|tip|danger|caution|error)(?=[\t ]|\r?\n|$)[^\r\n]*(?:\r?\n[\t ]*)*$/
-const SAFE_MARKDOWN_WINDOW_MARGIN = 1024
-const SAFE_MARKDOWN_WINDOW_OVERLAP = 16
-/**
- * Cached streaming safe-markdown transform, keyed by the md instance.
- *
- * The cache is owned by the TOP-LEVEL document stream: fragment parses
- * (e.g. `<details>` / custom html children, `__disableStructuredReuse`)
- * share the md instance and may overwrite the entry — benign, because the
- * transforms are deterministic functions of the source text, and the fast
- * path additionally guards with `mode` + `startsWith(previous.source)`, so
- * a stale/foreign entry either re-produces the identical transform or falls
- * back to a full-document transform.
- *
- * Memory: holds the full raw source + transformed output (~2-3x the document
- * size) per md instance for the instance lifetime; WeakMap-keyed so it is
- * collected with the instance. Cleared on the final auto-parse reset.
- */
-const safeMarkdownCache = new WeakMap<object, {
-  source: string
-  safeMarkdown: string
-  mode: string
-}>()
 
 interface ParseTimingMetrics {
   tokenCloneMs?: number
@@ -607,766 +545,6 @@ function shouldResetTopLevelStreamCacheForFinalAutoParse(md: MarkdownIt, options
     && typeof stream.reset === 'function'
 }
 
-function clearTolerantMathBoundaryStreamCache(md: MarkdownIt) {
-  tolerantMathBoundaryStreamCache.delete(md as unknown as object)
-}
-
-function createExplicitBracketMathContext(): ExplicitBracketMathContext {
-  return {
-    fenceChar: '',
-    fenceInBlockquote: false,
-    fenceInList: false,
-    fenceLen: 0,
-    fenceListIndent: 0,
-    inDollarMath: false,
-    inFence: false,
-    inMath: false,
-    listContentIndent: null,
-    dollarMathOpenOffset: null,
-    mathOpenOffset: null,
-  }
-}
-
-function cloneExplicitBracketMathContext(context: ExplicitBracketMathContext): ExplicitBracketMathContext {
-  return { ...context }
-}
-
-function setTolerantMathBoundaryStreamCache(
-  md: MarkdownIt,
-  source: string,
-  key: string | null,
-  explicitBracketMath: ExplicitBracketMathStreamState = scanExplicitBracketMathStreamState(source).state,
-) {
-  tolerantMathBoundaryStreamCache.set(md as unknown as object, {
-    explicitBracketMath,
-    source,
-    key,
-    pendingCandidate: key === null && mayContainTolerantMathBlockBoundaryOpener(source),
-  })
-}
-
-function sourceEndsWithSplitTolerantBoundaryPrefix(source: string) {
-  return source.endsWith('$') || source.endsWith('\\')
-}
-
-function sourceEndsWithCompleteTolerantBoundaryOpener(source: string) {
-  const lastLineStart = Math.max(source.lastIndexOf('\n') + 1, 0)
-  const lastLine = source.slice(lastLineStart).replace(/[\t ]+$/, '')
-  return TOLERANT_BOUNDARY_SPLIT_OPENERS.some(open => lastLine.endsWith(open))
-}
-
-function appendedChunkMayAffectTolerantMathBoundary(previousSource: string, appended: string) {
-  if (!appended)
-    return false
-
-  if (appended.includes('$$') || appended.includes('\\['))
-    return true
-
-  if (previousSource.endsWith('$') && appended[0] === '$')
-    return true
-
-  if (previousSource.endsWith('\\') && appended[0] === '[')
-    return true
-
-  // The opener may have arrived in the previous chunk:
-  //
-  //   "prefix $$" + "\na = 1"
-  //   "prefix \\[" + "\nx + y = z"
-  if (sourceEndsWithCompleteTolerantBoundaryOpener(previousSource) && /[\r\n]/.test(appended))
-    return true
-
-  return false
-}
-
-function isInsideOpenMarkdownFenceBeforeOffset(markdown: string, offset: number) {
-  let inFence = false
-  let fenceChar: '`' | '~' | '' = ''
-  let fenceLen = 0
-  let fenceInBlockquote = false
-  let fenceInList = false
-  let fenceListIndent = 0
-  let listContentIndent: number | null = null
-  let index = 0
-
-  while (index < offset) {
-    const newlineIndex = markdown.indexOf('\n', index)
-    const lineEnd = newlineIndex === -1 || newlineIndex >= offset ? offset : newlineIndex
-    const rawLine = markdown.slice(index, lineEnd)
-    const line = rawLine.endsWith('\r') ? rawLine.slice(0, -1) : rawLine
-    const lineIndent = getMarkdownIndent(line)
-    const listPrefix = stripMarkdownListPrefix(line)
-
-    if (inFence && fenceInBlockquote && line.trim() && stripMarkdownBlockquotePrefix(line) == null) {
-      inFence = false
-      fenceChar = ''
-      fenceLen = 0
-      fenceInBlockquote = false
-      fenceInList = false
-      fenceListIndent = 0
-    }
-
-    if (
-      inFence
-      && fenceInList
-      && line.trim()
-      && lineIndent.column < fenceListIndent
-      && !listPrefix
-    ) {
-      inFence = false
-      fenceChar = ''
-      fenceLen = 0
-      fenceInBlockquote = false
-      fenceInList = false
-      fenceListIndent = 0
-    }
-
-    if (listPrefix) {
-      listContentIndent = listPrefix.contentIndent
-    }
-    else if (line.trim() && listContentIndent != null && lineIndent.column < listContentIndent && !inFence) {
-      listContentIndent = null
-    }
-
-    const fenceMatch = matchMarkdownFenceMarker(line)
-    if (fenceMatch) {
-      if (inFence) {
-        if (
-          fenceMatch.markerChar === fenceChar
-          && fenceMatch.markerLen >= fenceLen
-          && /^\s*$/.test(fenceMatch.rest)
-        ) {
-          inFence = false
-          fenceChar = ''
-          fenceLen = 0
-          fenceInBlockquote = false
-          fenceInList = false
-          fenceListIndent = 0
-        }
-      }
-      else {
-        inFence = true
-        fenceChar = fenceMatch.markerChar
-        fenceLen = fenceMatch.markerLen
-        fenceInBlockquote = fenceMatch.inBlockquote
-        fenceInList = fenceMatch.inList
-          || (
-            listContentIndent != null
-            && !fenceMatch.inBlockquote
-            && lineIndent.column >= listContentIndent
-          )
-        fenceListIndent = fenceMatch.listIndent || listContentIndent || 0
-      }
-    }
-
-    if (newlineIndex === -1 || newlineIndex >= offset)
-      break
-    index = newlineIndex + 1
-  }
-
-  return inFence
-}
-
-function isInsideOpenStandardHtmlBlockBeforeOffset(markdown: string, offset: number) {
-  const isWs = (ch: string) => ch === ' ' || ch === '\t'
-  const isNameChar = (ch: string) => {
-    const c = ch.charCodeAt(0)
-    return (
-      (c >= 65 && c <= 90)
-      || (c >= 97 && c <= 122)
-      || (c >= 48 && c <= 57)
-      || ch === '_'
-      || ch === '-'
-      || ch === ':'
-    )
-  }
-  const parseLineStartTag = (line: string) => {
-    if (line[0] !== '<')
-      return null
-
-    let index = 1
-    while (index < line.length && isWs(line[index]))
-      index++
-
-    const closing = line[index] === '/'
-    if (closing) {
-      index++
-      while (index < line.length && isWs(line[index]))
-        index++
-    }
-
-    const nameStart = index
-    while (index < line.length && isNameChar(line[index]))
-      index++
-    if (index === nameStart)
-      return null
-
-    const tag = line.slice(nameStart, index).toLowerCase()
-    if (!STANDARD_BLOCK_HTML_TAGS.has(tag))
-      return null
-
-    const boundary = line[index]
-    if (boundary && boundary !== ' ' && boundary !== '\t' && boundary !== '>' && boundary !== '/')
-      return null
-
-    const tagEnd = findTagCloseIndexOutsideQuotes(line)
-    if (tagEnd === -1)
-      return null
-
-    let beforeEnd = tagEnd - 1
-    while (beforeEnd >= 0 && isWs(line[beforeEnd]))
-      beforeEnd--
-
-    return {
-      closing,
-      tag,
-      selfClosing: !closing && line[beforeEnd] === '/',
-      after: line.slice(tagEnd + 1),
-    }
-  }
-  const hasSameLineClose = (line: string, tag: string) => {
-    const lower = line.toLowerCase()
-    let index = 0
-    while (index < lower.length) {
-      const closeStart = lower.indexOf('</', index)
-      if (closeStart === -1)
-        return false
-      index = closeStart + 2
-      while (index < lower.length && isWs(lower[index]))
-        index++
-      if (lower.startsWith(tag, index)) {
-        const boundary = lower[index + tag.length]
-        if (!boundary || boundary === ' ' || boundary === '\t' || boundary === '>')
-          return true
-      }
-    }
-    return false
-  }
-
-  const stack: string[] = []
-  let inComment = false
-  let inDeclaration = false
-  let inProcessingInstruction = false
-  let index = 0
-
-  while (index < offset) {
-    const newlineIndex = markdown.indexOf('\n', index)
-    const lineEnd = newlineIndex === -1 || newlineIndex >= offset ? offset : newlineIndex
-    const rawLine = markdown.slice(index, lineEnd)
-    const line = rawLine.endsWith('\r') ? rawLine.slice(0, -1) : rawLine
-    const indent = consumeMarkdownIndent(line)
-    if (indent) {
-      const rest = line.slice(indent.index)
-      if (inComment) {
-        inComment = !rest.includes('-->')
-      }
-      else if (inDeclaration) {
-        inDeclaration = !rest.includes('>')
-      }
-      else if (inProcessingInstruction) {
-        inProcessingInstruction = !rest.includes('?>')
-      }
-      else if (rest.startsWith('<!--')) {
-        inComment = !rest.includes('-->')
-      }
-      else if (rest.startsWith('<?')) {
-        inProcessingInstruction = !rest.includes('?>')
-      }
-      else if (rest.startsWith('<!')) {
-        inDeclaration = !rest.includes('>')
-      }
-      else {
-        const tagInfo = parseLineStartTag(rest)
-        if (tagInfo) {
-          if (tagInfo.closing) {
-            for (let i = stack.length - 1; i >= 0; i--) {
-              if (stack[i] === tagInfo.tag) {
-                stack.length = i
-                break
-              }
-            }
-          }
-          else if (!tagInfo.selfClosing) {
-            if (!hasSameLineClose(tagInfo.after, tagInfo.tag))
-              stack.push(tagInfo.tag)
-          }
-        }
-      }
-    }
-
-    if (newlineIndex === -1 || newlineIndex >= offset)
-      break
-    index = newlineIndex + 1
-  }
-
-  return inComment || inDeclaration || inProcessingInstruction || stack.length > 0
-}
-
-function isInsideOpenCustomHtmlBlockBeforeOffset(markdown: string, offset: number, customHtmlTags?: readonly string[]) {
-  if (!customHtmlTags?.length)
-    return false
-
-  const tagSet = new Set(normalizeCustomHtmlTags(customHtmlTags))
-  if (!tagSet.size)
-    return false
-
-  const isNameChar = (ch: string) => {
-    const c = ch.charCodeAt(0)
-    return (
-      (c >= 65 && c <= 90)
-      || (c >= 97 && c <= 122)
-      || (c >= 48 && c <= 57)
-      || ch === '_'
-      || ch === '-'
-      || ch === ':'
-    )
-  }
-  const isWs = (ch: string) => ch === ' ' || ch === '\t'
-  const parseLineStartTag = (line: string) => {
-    if (line[0] !== '<')
-      return null
-
-    let index = 1
-    while (index < line.length && isWs(line[index]))
-      index++
-
-    const closing = line[index] === '/'
-    if (closing) {
-      index++
-      while (index < line.length && isWs(line[index]))
-        index++
-    }
-
-    const nameStart = index
-    while (index < line.length && isNameChar(line[index]))
-      index++
-    if (index === nameStart)
-      return null
-
-    const tag = line.slice(nameStart, index).toLowerCase()
-    if (!tagSet.has(tag))
-      return null
-
-    const boundary = line[index]
-    if (boundary && boundary !== ' ' && boundary !== '\t' && boundary !== '>' && boundary !== '/')
-      return null
-
-    const tagEnd = line.indexOf('>', index)
-    if (tagEnd === -1)
-      return null
-
-    let beforeEnd = tagEnd - 1
-    while (beforeEnd >= 0 && isWs(line[beforeEnd]))
-      beforeEnd--
-
-    return {
-      closing,
-      tag,
-      selfClosing: !closing && line[beforeEnd] === '/',
-      after: line.slice(tagEnd + 1),
-    }
-  }
-  const hasSameLineClose = (line: string, tag: string) => {
-    const lower = line.toLowerCase()
-    let index = 0
-    while (index < lower.length) {
-      const closeStart = lower.indexOf('</', index)
-      if (closeStart === -1)
-        return false
-      index = closeStart + 2
-      while (index < lower.length && isWs(lower[index]))
-        index++
-      if (lower.startsWith(tag, index)) {
-        const boundary = lower[index + tag.length]
-        if (!boundary || boundary === ' ' || boundary === '\t' || boundary === '>')
-          return true
-      }
-    }
-    return false
-  }
-
-  const stack: string[] = []
-  let index = 0
-
-  while (index < offset) {
-    const newlineIndex = markdown.indexOf('\n', index)
-    const lineEnd = newlineIndex === -1 || newlineIndex >= offset ? offset : newlineIndex
-    const rawLine = markdown.slice(index, lineEnd)
-    const line = rawLine.endsWith('\r') ? rawLine.slice(0, -1) : rawLine
-    const indent = consumeMarkdownIndent(line)
-    if (indent) {
-      const rest = line.slice(indent.index)
-      const tagInfo = parseLineStartTag(rest)
-      if (tagInfo) {
-        if (tagInfo.closing) {
-          for (let i = stack.length - 1; i >= 0; i--) {
-            if (stack[i] === tagInfo.tag) {
-              stack.length = i
-              break
-            }
-          }
-        }
-        else if (!tagInfo.selfClosing) {
-          if (!hasSameLineClose(tagInfo.after, tagInfo.tag))
-            stack.push(tagInfo.tag)
-        }
-      }
-    }
-
-    if (newlineIndex === -1 || newlineIndex >= offset)
-      break
-    index = newlineIndex + 1
-  }
-
-  return stack.length > 0
-}
-
-function getStreamingAdmonitionOpenTailReplacement(markdown: string, customHtmlTags?: readonly string[]) {
-  const match = STREAMING_ADMONITION_OPEN_RE.exec(markdown)
-  if (!match)
-    return null
-
-  const separator = match[1] ?? ''
-  const lineStart = match.index + separator.length
-  const lineEnd = markdown.indexOf('\n', lineStart)
-  const rawLine = markdown.slice(lineStart, lineEnd === -1 ? markdown.length : lineEnd)
-  const line = rawLine.endsWith('\r') ? rawLine.slice(0, -1) : rawLine
-  if (!consumeMarkdownIndent(line))
-    return null
-
-  if (isInsideOpenMarkdownFenceBeforeOffset(markdown, lineStart))
-    return null
-
-  if (isInsideOpenStandardHtmlBlockBeforeOffset(markdown, lineStart))
-    return null
-
-  if (isInsideOpenCustomHtmlBlockBeforeOffset(markdown, lineStart, customHtmlTags))
-    return null
-
-  return `${markdown.slice(0, match.index)}${separator}`
-}
-
-function resetExplicitBracketFenceContext(context: ExplicitBracketMathContext) {
-  context.inFence = false
-  context.fenceChar = ''
-  context.fenceLen = 0
-  context.fenceInBlockquote = false
-  context.fenceInList = false
-  context.fenceListIndent = 0
-}
-
-function scanLineForExplicitBracketMathState(
-  line: string,
-  context: ExplicitBracketMathContext,
-  lineStart: number,
-  appendStart: number | null,
-  openAtAppendStart: boolean,
-) {
-  let index = 0
-  let closedOpenMath = false
-
-  while (index < line.length) {
-    const sourceIndex = index
-
-    if (context.inMath) {
-      if (line.startsWith('\\]', index) && !isEscapedDelimiterAt(line, sourceIndex)) {
-        if (appendStart != null && openAtAppendStart && lineStart + index + 2 > appendStart)
-          closedOpenMath = true
-        context.inMath = false
-        context.mathOpenOffset = null
-        index += 2
-        continue
-      }
-
-      index++
-      continue
-    }
-
-    if (context.inDollarMath) {
-      if (line.startsWith('$$', index) && !isEscapedDelimiterAt(line, sourceIndex)) {
-        if (appendStart != null && openAtAppendStart && lineStart + index + 2 > appendStart)
-          closedOpenMath = true
-        context.inDollarMath = false
-        context.dollarMathOpenOffset = null
-        index += 2
-        continue
-      }
-
-      index++
-      continue
-    }
-
-    if (line[index] === '`' && !isEscapedDelimiterAt(line, sourceIndex)) {
-      const markerLen = countRepeatedChar(line, index, '`')
-      const closeIndex = findCodeSpanCloseIndex(line, index + markerLen, markerLen)
-      if (closeIndex === -1)
-        break
-      index = closeIndex + markerLen
-      continue
-    }
-
-    if (line.startsWith('\\[', index) && !isEscapedDelimiterAt(line, sourceIndex)) {
-      context.inMath = true
-      context.mathOpenOffset = lineStart + index
-      index += 2
-      continue
-    }
-
-    if (line.startsWith('$$', index) && !isEscapedDelimiterAt(line, sourceIndex)) {
-      context.inDollarMath = true
-      context.dollarMathOpenOffset = lineStart + index
-      index += 2
-      continue
-    }
-
-    index++
-  }
-
-  return closedOpenMath
-}
-
-function stripPendingExplicitMathTail(markdown: string, md: MarkdownIt) {
-  if (!hasMarkstreamMathPlugin(md))
-    return markdown
-
-  const owner = md as unknown as object
-  const previous = pendingExplicitMathTailCache.get(owner)
-  const state = previous?.source === markdown
-    ? previous.state
-    : previous && markdown.startsWith(previous.source)
-      ? updateExplicitBracketMathStreamState(
-        previous.state,
-        markdown.slice(previous.source.length),
-        previous.source.length - previous.state.lineBuffer.length,
-      ).state
-      : scanExplicitBracketMathStreamState(markdown).state
-  pendingExplicitMathTailCache.set(owner, { source: markdown, state })
-
-  const { context } = state
-  const openOffset = context.inMath
-    ? context.mathOpenOffset
-    : context.inDollarMath ? context.dollarMathOpenOffset : null
-  if (openOffset == null)
-    return markdown
-
-  const content = markdown.slice(openOffset + 2)
-  const lineStart = markdown.lastIndexOf('\n', openOffset - 1) + 1
-  const startsBlockLine = markdown.slice(lineStart, openOffset).trim() === ''
-  if (!startsBlockLine && !/^\r?\n/.test(content))
-    return markdown
-  if (/^\s*!\[/.test(content))
-    return markdown
-
-  const stripped = content.trim()
-  const weakSingleVariable = /^(?:[a-z]|pi)$/i.test(stripped)
-  if (isMathLike(content) && !weakSingleVariable)
-    return markdown
-
-  return markdown.slice(0, openOffset)
-}
-
-function scanExplicitBracketMathLine(
-  line: string,
-  context: ExplicitBracketMathContext,
-  lineStart: number,
-  appendStart: number | null,
-  openAtAppendStart: boolean,
-) {
-  const lineIndent = getMarkdownIndent(line)
-  const listPrefix = stripMarkdownListPrefix(line)
-
-  if (context.inFence && context.fenceInBlockquote && line.trim() && stripMarkdownBlockquotePrefix(line) == null)
-    resetExplicitBracketFenceContext(context)
-
-  if (
-    context.inFence
-    && context.fenceInList
-    && line.trim()
-    && lineIndent.column < context.fenceListIndent
-    && !listPrefix
-  ) {
-    resetExplicitBracketFenceContext(context)
-  }
-
-  if (listPrefix) {
-    context.listContentIndent = listPrefix.contentIndent
-  }
-  else if (
-    line.trim()
-    && context.listContentIndent != null
-    && lineIndent.column < context.listContentIndent
-    && !context.inFence
-  ) {
-    context.listContentIndent = null
-  }
-
-  if (!context.inMath && !context.inDollarMath) {
-    const fenceMatch = matchMarkdownFenceMarker(line)
-    if (fenceMatch) {
-      if (context.inFence) {
-        if (
-          fenceMatch.markerChar === context.fenceChar
-          && fenceMatch.markerLen >= context.fenceLen
-          && /^\s*$/.test(fenceMatch.rest)
-        ) {
-          resetExplicitBracketFenceContext(context)
-        }
-      }
-      else {
-        context.inFence = true
-        context.fenceChar = fenceMatch.markerChar
-        context.fenceLen = fenceMatch.markerLen
-        context.fenceInBlockquote = fenceMatch.inBlockquote
-        context.fenceInList = fenceMatch.inList
-          || (
-            context.listContentIndent != null
-            && !fenceMatch.inBlockquote
-            && lineIndent.column >= context.listContentIndent
-          )
-        context.fenceListIndent = fenceMatch.listIndent || context.listContentIndent || 0
-      }
-    }
-    else if (!context.inFence) {
-      return scanLineForExplicitBracketMathState(line, context, lineStart, appendStart, openAtAppendStart)
-    }
-  }
-  else {
-    return scanLineForExplicitBracketMathState(line, context, lineStart, appendStart, openAtAppendStart)
-  }
-
-  return false
-}
-
-function scanExplicitBracketMathStreamState(
-  source: string,
-  initialContext: ExplicitBracketMathContext = createExplicitBracketMathContext(),
-  appendStart: number | null = null,
-  openAtAppendStart = false,
-  sourceOffset = 0,
-) {
-  const context = cloneExplicitBracketMathContext(initialContext)
-  let committedContext = cloneExplicitBracketMathContext(initialContext)
-  let lineBuffer = ''
-  let closedOpenMath = false
-  let index = 0
-
-  while (index < source.length) {
-    const newlineIndex = source.indexOf('\n', index)
-    const hasNewline = newlineIndex !== -1
-    const lineEnd = hasNewline && newlineIndex > index && source[newlineIndex - 1] === '\r'
-      ? newlineIndex - 1
-      : hasNewline ? newlineIndex : source.length
-    const line = source.slice(index, lineEnd)
-
-    if (scanExplicitBracketMathLine(line, context, sourceOffset + index, appendStart, openAtAppendStart))
-      closedOpenMath = true
-
-    if (hasNewline) {
-      committedContext = cloneExplicitBracketMathContext(context)
-      lineBuffer = ''
-    }
-    else {
-      lineBuffer = line
-    }
-
-    index = hasNewline ? newlineIndex + 1 : source.length
-  }
-
-  return {
-    closedOpenMath,
-    state: {
-      committedContext,
-      context,
-      lineBuffer,
-    },
-  }
-}
-
-function updateExplicitBracketMathStreamState(
-  previous: ExplicitBracketMathStreamState,
-  appended: string,
-  lineBufferStartOffset = 0,
-) {
-  if (
-    appended
-    && !previous.context.inMath
-    && !previous.context.inDollarMath
-    && !previous.context.inFence
-    && !previous.committedContext.inFence
-    && !/[\\$`~\r\n]/.test(appended)
-    && !(previous.lineBuffer.endsWith('\\') && (appended[0] === '[' || appended[0] === ']'))
-  ) {
-    return {
-      closedOpenMath: false,
-      state: {
-        committedContext: cloneExplicitBracketMathContext(previous.committedContext),
-        context: cloneExplicitBracketMathContext(previous.context),
-        lineBuffer: previous.lineBuffer + appended,
-      },
-    }
-  }
-
-  return scanExplicitBracketMathStreamState(
-    previous.lineBuffer + appended,
-    previous.committedContext,
-    lineBufferStartOffset + previous.lineBuffer.length,
-    previous.context.inMath || previous.context.inDollarMath,
-    lineBufferStartOffset,
-  )
-}
-
-function syncTolerantMathBoundaryStreamCache(md: MarkdownIt, source: string) {
-  if (!hasMarkstreamMathPlugin(md))
-    return
-
-  const stream = md.stream
-  if (typeof stream?.reset !== 'function')
-    return
-
-  const owner = md as unknown as object
-  const previous = tolerantMathBoundaryStreamCache.get(owner)
-
-  if (previous?.source === source)
-    return
-
-  const sourceExtendsPrevious = previous ? source.startsWith(previous.source) : false
-  const appended = sourceExtendsPrevious && previous ? source.slice(previous.source.length) : ''
-  const explicitBracketMathUpdate = sourceExtendsPrevious && previous
-    ? updateExplicitBracketMathStreamState(
-        previous.explicitBracketMath,
-        appended,
-        previous.source.length - previous.explicitBracketMath.lineBuffer.length,
-      )
-    : scanExplicitBracketMathStreamState(source)
-  const nextExplicitBracketMath = explicitBracketMathUpdate.state
-  const completesExplicitBracketMathClose = sourceExtendsPrevious && previous
-    ? explicitBracketMathUpdate.closedOpenMath
-    : false
-
-  if (previous && sourceExtendsPrevious) {
-    if (
-      previous.key === null
-      && previous.pendingCandidate === false
-      && !completesExplicitBracketMathClose
-      && !appendedChunkMayAffectTolerantMathBoundary(previous.source, appended)
-      && !sourceEndsWithSplitTolerantBoundaryPrefix(source)
-    ) {
-      previous.source = source
-      previous.explicitBracketMath = nextExplicitBracketMath
-      return
-    }
-  }
-
-  const nextKey = getTolerantMathBlockBoundaryStreamKey(source)
-  const sourceWasReplaced = previous ? !sourceExtendsPrevious : false
-
-  if (previous && (sourceWasReplaced || previous.key !== nextKey || completesExplicitBracketMathClose))
-    stream.reset()
-  else if (!previous && nextKey)
-    stream.reset()
-
-  setTolerantMathBoundaryStreamCache(md, source, nextKey, nextExplicitBracketMath)
-}
-
 function shouldCloneTopLevelStreamTokens(options: ParseOptions) {
   return typeof options.preTransformTokens === 'function'
     || typeof options.postTransformTokens === 'function'
@@ -1479,11 +657,6 @@ function shouldFallbackDuplicateTolerantMathStreamTokens(
   return hasMarkstreamMathPlugin(md)
     && mayContainTolerantMathBlockBoundaryOpener(source)
     && hasAdjacentDuplicateParagraphTokenTriplet(tokens)
-}
-
-function shouldUseSyncParseForPendingTolerantMathBoundary(md: MarkdownIt) {
-  const cache = tolerantMathBoundaryStreamCache.get(md as unknown as object)
-  return typeof cache?.key === 'string' && cache.key.startsWith('pending:')
 }
 
 function parseTopLevelTokens(
@@ -2424,1395 +1597,6 @@ function mergeSplitTopLevelHtmlBlocks(nodes: ParsedNode[], final: boolean, sourc
   return merged
 }
 
-function stripDanglingHtmlLikeTail(markdown: string) {
-  const isWs = (ch: string) => ch === ' ' || ch === '\t' || ch === '\n' || ch === '\r'
-
-  const isLikelyHtmlTagPrefix = (tail: string) => {
-    // Deterministic scanner (avoids ReDoS/backtracking regexes).
-    // Accepts prefixes like "<think", "</think", "<div class", "<a href=\"x"
-    // and treats them as "HTML-ish" tails that can be stripped in streaming mode.
-    if (!tail || tail[0] !== '<')
-      return false
-    if (tail.includes('>'))
-      return false
-
-    let i = 1
-    // "< " is likely comparison ("x < y"), not a tag
-    if (i < tail.length && isWs(tail[i]))
-      return false
-
-    if (tail.startsWith('<!--') || tail.startsWith('<?') || tail.startsWith('<!'))
-      return false
-
-    if (tail[i] === '/') {
-      i++
-      // "</ " isn't a tag start
-      if (i < tail.length && isWs(tail[i]))
-        return false
-    }
-
-    const isAlpha = (ch: string) => {
-      const c = ch.charCodeAt(0)
-      return (c >= 65 && c <= 90) || (c >= 97 && c <= 122)
-    }
-    const isDigit = (ch: string) => {
-      const c = ch.charCodeAt(0)
-      return c >= 48 && c <= 57
-    }
-    const isNameStart = (ch: string) => ch === '!' || isAlpha(ch)
-    const isNameChar = (ch: string) => isAlpha(ch) || isDigit(ch) || ch === ':' || ch === '-'
-    const isAttrStart = (ch: string) => isAlpha(ch) || isDigit(ch) || ch === '_' || ch === '.' || ch === ':' || ch === '-'
-    const isAttrChar = isAttrStart
-
-    if (i >= tail.length || !isNameStart(tail[i]))
-      return false
-
-    // tag name
-    i++
-    while (i < tail.length && isNameChar(tail[i]))
-      i++
-
-    while (i < tail.length) {
-      // trailing whitespace ok
-      while (i < tail.length && isWs(tail[i]))
-        i++
-      if (i >= tail.length)
-        return true
-
-      // allow self-closing slash at end (e.g. "<br/")
-      if (tail[i] === '/') {
-        i++
-        while (i < tail.length && isWs(tail[i]))
-          i++
-        return i >= tail.length
-      }
-
-      // attribute name
-      if (!isAttrStart(tail[i]))
-        return false
-      i++
-      while (i < tail.length && isAttrChar(tail[i]))
-        i++
-
-      while (i < tail.length && isWs(tail[i]))
-        i++
-
-      if (i < tail.length && tail[i] === '=') {
-        i++
-        while (i < tail.length && isWs(tail[i]))
-          i++
-        if (i >= tail.length)
-          return true // incomplete value
-
-        const quote = tail[i]
-        if (quote === '"' || quote === '\'') {
-          i++
-          while (i < tail.length && tail[i] !== quote)
-            i++
-          // If we don't see the closing quote (tail ends), it's still a tag prefix
-          if (i >= tail.length)
-            return true
-          i++ // consume closing quote
-        }
-        else {
-          // unquoted value: scan until whitespace or forbidden delimiters
-          while (i < tail.length) {
-            const ch = tail[i]
-            if (isWs(ch) || ch === '<' || ch === '>' || ch === '"' || ch === '\'' || ch === '`')
-              break
-            i++
-          }
-          if (i >= tail.length)
-            return true // incomplete unquoted value
-        }
-      }
-      // else: boolean attr, continue
-    }
-
-    return true
-  }
-
-  // Delegate to the full fence scanner used elsewhere: the previous local
-  // scanner only recognized direct and blockquote-prefixed fences, so a
-  // fence nested inside a list item (`- ```html` / `  <div`) was invisible
-  // and the incomplete `<div` tail got truncated from the code content on
-  // every non-final commit. `isInsideOpenMarkdownFenceBeforeOffset` also
-  // handles list/blockquote fence exit conditions (de-dent ends the fence).
-  const isInsideFencedCodeBlock = (src: string, pos: number) =>
-    isInsideOpenMarkdownFenceBeforeOffset(src, pos)
-
-  // In streaming mode it's common to have an incomplete HTML-ish fragment at
-  // the very end of the current buffer (e.g. '<fo' or '</think'). Letting it
-  // reach markdown-it can produce visible mid-state text nodes. We only strip
-  // the *tail* when there is no closing '>' anywhere after the last '<'.
-  const s = String(markdown ?? '')
-  const lastLt = s.lastIndexOf('<')
-  if (lastLt === -1)
-    return s
-  if (isInsideFencedCodeBlock(s, lastLt))
-    return s
-
-  // Only treat it as an HTML-ish tail when "<" looks like a tag start.
-  // This avoids truncating normal text/math like "y_{<i}" or "x < y".
-  if (lastLt > 0) {
-    const prev = s[lastLt - 1]
-    const prevIsWs = prev === ' ' || prev === '\t' || prev === '\n' || prev === '\r'
-    // Some stream transports escape newlines as "\\n" / "\\r\\n". Treat those
-    // sequences as line boundaries too.
-    const prev2 = s[lastLt - 2]
-    const prevLooksLikeEscapedNewline = (prev === 'n' || prev === 'r') && prev2 === '\\'
-    if (!prevIsWs && !prevLooksLikeEscapedNewline)
-      return s
-  }
-
-  const tail = s.slice(lastLt)
-  if (tail.includes('>'))
-    return s
-  // If the char after '<' is whitespace, it's more likely a comparison ("x < y")
-  // than a tag start ("<div").
-  if (tail.length > 1 && (tail[1] === ' ' || tail[1] === '\t' || tail[1] === '\n' || tail[1] === '\r'))
-    return s
-
-  if (!isLikelyHtmlTagPrefix(tail))
-    return s
-  return s.slice(0, lastLt)
-}
-
-function ensureBlankLineBeforeInlineMultilineCustomHtmlBlocks(markdown: string, tags: string[]) {
-  if (!markdown || !tags.length)
-    return markdown
-
-  const tagSet = new Set(tags.map(t => String(t ?? '').toLowerCase()).filter(Boolean))
-  if (!tagSet.size)
-    return markdown
-
-  const isIndentWs = (ch: string) => ch === ' ' || ch === '\t'
-  const isNameChar = (ch: string) => {
-    const c = ch.charCodeAt(0)
-    return (
-      (c >= 65 && c <= 90) // A-Z
-      || (c >= 97 && c <= 122) // a-z
-      || (c >= 48 && c <= 57) // 0-9
-      || ch === '_'
-      || ch === '-'
-      || ch === ':'
-    )
-  }
-
-  const isIndentedCodeLine = (line: string) => {
-    if (!line)
-      return false
-    if (line[0] === '\t')
-      return true
-    let spaces = 0
-    for (let i = 0; i < line.length; i++) {
-      const ch = line[i]
-      if (ch === ' ') {
-        spaces++
-        if (spaces >= 4)
-          return true
-        continue
-      }
-      if (ch === '\t')
-        return true
-      break
-    }
-    return false
-  }
-
-  const findTagCloseIndexOutsideQuotes = (input: string) => {
-    let inSingle = false
-    let inDouble = false
-    for (let i = 0; i < input.length; i++) {
-      const ch = input[i]
-      if (ch === '\\') {
-        i++
-        continue
-      }
-      if (!inDouble && ch === '\'') {
-        inSingle = !inSingle
-        continue
-      }
-      if (!inSingle && ch === '"') {
-        inDouble = !inDouble
-        continue
-      }
-      if (!inSingle && !inDouble && ch === '>')
-        return i
-    }
-    return -1
-  }
-
-  const parseFenceMarker = (line: string) => {
-    let i = 0
-    while (i < line.length && isIndentWs(line[i])) i++
-    const ch = line[i]
-    if (ch !== '`' && ch !== '~')
-      return null
-    let j = i
-    while (j < line.length && line[j] === ch) j++
-    const len = j - i
-    if (len < 3)
-      return null
-    return { markerChar: ch as '`' | '~', markerLen: len, rest: line.slice(j) }
-  }
-
-  const findInlineCustomBlockSplitIndex = (line: string, lineStart: number) => {
-    if (isIndentedCodeLine(line))
-      return -1
-
-    const trimmed = line.replace(/^[ \t]+/, '')
-    if (!trimmed || trimmed.startsWith('>') || trimmed.startsWith('|') || /^(?:[*+-]|\d+[.)])[\t ]+/.test(trimmed))
-      return -1
-
-    let hasRenderablePrefix = false
-    let i = 0
-    while (i < line.length) {
-      const ch = line[i]
-      if (ch !== '<') {
-        if (!isIndentWs(ch))
-          hasRenderablePrefix = true
-        i++
-        continue
-      }
-
-      const closeIdxRel = findTagCloseIndexOutsideQuotes(line.slice(i))
-      if (closeIdxRel === -1) {
-        hasRenderablePrefix = true
-        i++
-        continue
-      }
-
-      const tagSlice = line.slice(i, i + closeIdxRel + 1)
-      let cursor = 1
-      while (cursor < tagSlice.length && isIndentWs(tagSlice[cursor])) cursor++
-      if (cursor >= tagSlice.length) {
-        hasRenderablePrefix = true
-        i++
-        continue
-      }
-
-      const marker = tagSlice[cursor]
-      if (marker === '!' || marker === '?') {
-        hasRenderablePrefix = true
-        i += closeIdxRel + 1
-        continue
-      }
-
-      const isClosing = marker === '/'
-      if (isClosing) {
-        hasRenderablePrefix = true
-        i += closeIdxRel + 1
-        continue
-      }
-
-      const nameStart = cursor
-      while (cursor < tagSlice.length && isNameChar(tagSlice[cursor])) cursor++
-      if (cursor === nameStart) {
-        hasRenderablePrefix = true
-        i++
-        continue
-      }
-
-      const tagName = tagSlice.slice(nameStart, cursor).toLowerCase()
-      const boundary = tagSlice[cursor]
-      if (boundary && boundary !== ' ' && boundary !== '\t' && boundary !== '>' && boundary !== '/') {
-        hasRenderablePrefix = true
-        i++
-        continue
-      }
-
-      const sameLineCloseRe = new RegExp(String.raw`<\s*\/\s*${tagName}\s*>`, 'i')
-      const selfClosing = /\/\s*>$/.test(tagSlice)
-      const closesOnSameLine = sameLineCloseRe.test(line.slice(i + closeIdxRel + 1))
-      const closesLater = sameLineCloseRe.test(markdown.slice(lineStart + i + closeIdxRel + 1))
-      const continuesOnLaterLine = /[\r\n]/.test(markdown.slice(lineStart + i + closeIdxRel + 1))
-
-      if (hasRenderablePrefix && tagSet.has(tagName) && !selfClosing && !closesOnSameLine && (closesLater || continuesOnLaterLine))
-        return i
-
-      hasRenderablePrefix = true
-      i += closeIdxRel + 1
-    }
-
-    return -1
-  }
-
-  let inFence = false
-  let fenceChar: '`' | '~' | '' = ''
-  let fenceLen = 0
-  let out = ''
-  let idx = 0
-
-  while (idx < markdown.length) {
-    const nl = markdown.indexOf('\n', idx)
-    const hasNl = nl !== -1
-    const isCrlf = hasNl && nl > idx && markdown[nl - 1] === '\r'
-    const lineEnd = hasNl ? (isCrlf ? nl - 1 : nl) : markdown.length
-    const line = markdown.slice(idx, lineEnd)
-    const newline = hasNl ? (isCrlf ? '\r\n' : '\n') : ''
-
-    const fenceMatch = parseFenceMarker(line)
-    let nextLine = line
-    if (!inFence && !fenceMatch) {
-      const splitAt = findInlineCustomBlockSplitIndex(line, idx)
-      if (splitAt !== -1) {
-        const separator = newline || '\n'
-        const before = line.slice(0, splitAt).replace(/[ \t]+$/, '')
-        const after = line.slice(splitAt).replace(/^[ \t]+/, '')
-        nextLine = `${before}${separator}${separator}${after}`
-      }
-    }
-
-    out += nextLine
-    out += newline
-
-    if (fenceMatch) {
-      if (inFence) {
-        if (fenceMatch.markerChar === fenceChar && fenceMatch.markerLen >= fenceLen) {
-          if (/^\s*$/.test(fenceMatch.rest)) {
-            inFence = false
-            fenceChar = ''
-            fenceLen = 0
-          }
-        }
-      }
-      else {
-        inFence = true
-        fenceChar = fenceMatch.markerChar
-        fenceLen = fenceMatch.markerLen
-      }
-    }
-
-    idx = hasNl ? nl + 1 : markdown.length
-  }
-
-  return out
-}
-
-function normalizeCustomHtmlOpeningTagSameLine(markdown: string, tags: string[]) {
-  if (!markdown || !tags.length)
-    return markdown
-
-  const tagSet = new Set(tags.map(t => String(t ?? '').toLowerCase()))
-  if (!tagSet.size)
-    return markdown
-
-  const isIndentWs = (ch: string) => ch === ' ' || ch === '\t'
-  const isNameChar = (ch: string) => {
-    const c = ch.charCodeAt(0)
-    return (
-      (c >= 65 && c <= 90) // A-Z
-      || (c >= 97 && c <= 122) // a-z
-      || (c >= 48 && c <= 57) // 0-9
-      || ch === '_'
-      || ch === '-'
-    )
-  }
-
-  const trimStartIndentWs = (s: string) => {
-    let i = 0
-    while (i < s.length && isIndentWs(s[i])) i++
-    return s.slice(i)
-  }
-
-  const findTagCloseIndexOutsideQuotes = (input: string) => {
-    let inSingle = false
-    let inDouble = false
-    for (let i = 0; i < input.length; i++) {
-      const ch = input[i]
-      if (ch === '\\') {
-        i++
-        continue
-      }
-      if (!inDouble && ch === '\'') {
-        inSingle = !inSingle
-        continue
-      }
-      if (!inSingle && ch === '"') {
-        inDouble = !inDouble
-        continue
-      }
-      if (!inSingle && !inDouble && ch === '>')
-        return i
-    }
-    return -1
-  }
-
-  const hasClosingTagOnLine = (line: string, from: number, tag: string) => {
-    const lowerTag = tag.toLowerCase()
-    let pos = line.indexOf('<', from)
-    while (pos !== -1) {
-      let i = pos + 1
-      while (i < line.length && isIndentWs(line[i])) i++
-      if (i >= line.length || line[i] !== '/') {
-        pos = line.indexOf('<', pos + 1)
-        continue
-      }
-      i++
-      while (i < line.length && isIndentWs(line[i])) i++
-      if (i + lowerTag.length > line.length) {
-        pos = line.indexOf('<', pos + 1)
-        continue
-      }
-
-      // Case-insensitive match for the closing tag name.
-      let matched = true
-      for (let j = 0; j < lowerTag.length; j++) {
-        const ch = line[i + j]
-        const lc = ch >= 'A' && ch <= 'Z' ? String.fromCharCode(ch.charCodeAt(0) + 32) : ch
-        if (lc !== lowerTag[j]) {
-          matched = false
-          break
-        }
-      }
-      if (!matched) {
-        pos = line.indexOf('<', pos + 1)
-        continue
-      }
-
-      let k = i + lowerTag.length
-      // Ensure exact tag name (no extra name characters).
-      if (k < line.length && isNameChar(line[k])) {
-        pos = line.indexOf('<', pos + 1)
-        continue
-      }
-      while (k < line.length && isIndentWs(line[k])) k++
-      if (k < line.length && line[k] === '>')
-        return true
-
-      pos = line.indexOf('<', pos + 1)
-    }
-    return false
-  }
-
-  const normalizeLine = (line: string) => {
-    let i = 0
-    while (i < line.length && isIndentWs(line[i])) i++
-    if (i >= line.length || line[i] !== '<')
-      return line
-
-    i++
-    while (i < line.length && isIndentWs(line[i])) i++
-    if (i >= line.length || line[i] === '/')
-      return line
-
-    const nameStart = i
-    while (i < line.length && isNameChar(line[i])) i++
-    if (i === nameStart)
-      return line
-
-    const tagName = line.slice(nameStart, i).toLowerCase()
-    if (!tagSet.has(tagName))
-      return line
-
-    const gtRel = findTagCloseIndexOutsideQuotes(line.slice(i))
-    if (gtRel === -1)
-      return line
-    const gt = i + gtRel
-
-    if (hasClosingTagOnLine(line, gt + 1, tagName))
-      return line
-
-    const rest = trimStartIndentWs(line.slice(gt + 1))
-    if (!rest)
-      return line
-
-    return `${line.slice(0, gt + 1)}\n${rest}`
-  }
-
-  let out = ''
-  let idx = 0
-  while (idx < markdown.length) {
-    const nl = markdown.indexOf('\n', idx)
-    if (nl === -1) {
-      out += normalizeLine(markdown.slice(idx))
-      break
-    }
-
-    const isCrlf = nl > idx && markdown[nl - 1] === '\r'
-    const lineEnd = isCrlf ? nl - 1 : nl
-    const line = markdown.slice(idx, lineEnd)
-    out += normalizeLine(line)
-    out += isCrlf ? '\r\n' : '\n'
-    idx = nl + 1
-  }
-
-  return out
-}
-
-function ensureBlankLineAfterCustomHtmlCloseBeforeBlockMarkerSameLine(markdown: string, tags: string[]) {
-  if (!markdown || !tags.length)
-    return markdown
-
-  const tagSet = new Set(tags.map(t => String(t ?? '').toLowerCase()))
-  if (!tagSet.size)
-    return markdown
-
-  const isIndentWs = (ch: string) => ch === ' ' || ch === '\t'
-
-  const parseBlockquotePrefix = (rawLine: string) => {
-    let i = 0
-    let saw = false
-    let prefixEnd = 0
-
-    while (i < rawLine.length) {
-      while (i < rawLine.length && isIndentWs(rawLine[i])) i++
-      if (i >= rawLine.length || rawLine[i] !== '>')
-        break
-      saw = true
-      i++
-      while (i < rawLine.length && isIndentWs(rawLine[i])) i++
-      prefixEnd = i
-    }
-
-    if (!saw)
-      return null
-
-    const prefix = rawLine.slice(0, prefixEnd)
-    return { prefix, content: rawLine.slice(prefixEnd) }
-  }
-
-  const parseFenceMarker = (line: string) => {
-    let i = 0
-    while (i < line.length && isIndentWs(line[i])) i++
-    const ch = line[i]
-    if (ch !== '`' && ch !== '~')
-      return null
-    let j = i
-    while (j < line.length && line[j] === ch) j++
-    const len = j - i
-    if (len < 3)
-      return null
-    return { markerChar: ch as '`' | '~', markerLen: len, rest: line.slice(j) }
-  }
-
-  const closeTagRes = Array.from(tagSet).map((tag) => {
-    // Insert a blank line after the close tag when the remaining same-line
-    // content begins with a block-level marker (e.g. "## ", "- ", "> ", "```", "|", "$$", ":::").
-    //
-    // Note: this is intentionally conservative and only targets constructs that
-    // require line-start to be recognized by markdown-it.
-    const blockMarkerLookahead = '(?=[\\t ]*(?:#{1,6}[\\t ]+|>|(?:[*+-]|\\d+[.)])[\\t ]+|(?:`{3,}|~{3,})|\\||\\$\\$|:{3,}|\\[\\^[^\\]]+\\]:|(?:-{3,}|\\*{3,}|_{3,})))'
-    return new RegExp(String.raw`(<\s*\/\s*${tag}\s*>)${blockMarkerLookahead}`, 'gi')
-  })
-
-  let inFence = false
-  let fenceChar: '`' | '~' | '' = ''
-  let fenceLen = 0
-
-  let out = ''
-  let idx = 0
-
-  while (idx < markdown.length) {
-    const nl = markdown.indexOf('\n', idx)
-    const hasNl = nl !== -1
-    const isCrlf = hasNl && nl > idx && markdown[nl - 1] === '\r'
-    const lineEnd = hasNl ? (isCrlf ? nl - 1 : nl) : markdown.length
-    const rawLine = markdown.slice(idx, lineEnd)
-    const newline = hasNl ? (isCrlf ? '\r\n' : '\n') : ''
-
-    const bq = parseBlockquotePrefix(rawLine)
-    const prefix = bq?.prefix ?? ''
-    const contentLine = bq?.content ?? rawLine
-
-    // Track fenced code blocks (including those nested in blockquotes) so we
-    // don't mutate their contents.
-    const fenceMatch = parseFenceMarker(contentLine)
-    if (fenceMatch) {
-      if (inFence) {
-        if (fenceMatch.markerChar === fenceChar && fenceMatch.markerLen >= fenceLen) {
-          if (/^\s*$/.test(fenceMatch.rest)) {
-            inFence = false
-            fenceChar = ''
-            fenceLen = 0
-          }
-        }
-      }
-      else {
-        inFence = true
-        fenceChar = fenceMatch.markerChar
-        fenceLen = fenceMatch.markerLen
-      }
-    }
-
-    let nextContent = contentLine
-    if (!inFence && nextContent.includes('</')) {
-      for (const re of closeTagRes) {
-        nextContent = nextContent.replace(re, (match, closeTag: string, offset: number, src: string) => {
-          // Inside table rows like:
-          //   | A | <my_component></my_component>## heading-like |
-          // do not inject blank lines after the closing tag, otherwise the row
-          // gets split and table parsing breaks after this custom cell.
-          const lineTrimmed = src.replace(/^[\t ]+/, '')
-          if (lineTrimmed.startsWith('|'))
-            return match
-
-          const before = src.slice(0, offset).replace(/^[\t ]+/, '')
-          // Keep same-line boundary splitting conservative:
-          // only split when the line starts with the custom tag block itself,
-          // or when the close tag is at line start (e.g. "</tag>## heading").
-          // This avoids breaking list/blockquote/paragraph inline contexts like:
-          // "- text <my_component></my_component>## h"
-          // "> text <my_component></my_component>- item"
-          // "text <my_component></my_component>## h"
-          if (before.length > 0) {
-            const closeTagName = closeTag.match(/^<\s*\/\s*([A-Z][\w:-]*)/i)?.[1]?.toLowerCase() ?? ''
-            const openTagName = before.match(/^<\s*([A-Z][\w:-]*)/i)?.[1]?.toLowerCase() ?? ''
-            if (!closeTagName || !openTagName || closeTagName !== openTagName)
-              return match
-          }
-
-          return `${closeTag}\n\n`
-        })
-      }
-    }
-
-    if (prefix) {
-      const withPrefix = prefix + nextContent.split('\n').join(`\n${prefix}`)
-      out += withPrefix
-    }
-    else {
-      out += nextContent
-    }
-
-    out += newline
-    idx = hasNl ? nl + 1 : markdown.length
-  }
-
-  return out
-}
-
-function ensureBlankLineBeforeCustomHtmlBlocks(markdown: string, tags: string[]) {
-  if (!markdown || !tags.length)
-    return markdown
-
-  const tagSet = new Set(tags.map(t => String(t ?? '').toLowerCase()))
-  if (!tagSet.size)
-    return markdown
-
-  const isIndentWs = (ch: string) => ch === ' ' || ch === '\t'
-  const isIndentedCodeLine = (line: string) => {
-    if (!line)
-      return false
-    if (line[0] === '\t')
-      return true
-    let spaces = 0
-    for (let i = 0; i < line.length; i++) {
-      const ch = line[i]
-      if (ch === ' ') {
-        spaces++
-        if (spaces >= 4)
-          return true
-        continue
-      }
-      if (ch === '\t')
-        return true
-      break
-    }
-    return false
-  }
-  const isNameChar = (ch: string) => {
-    const c = ch.charCodeAt(0)
-    return (
-      (c >= 65 && c <= 90) // A-Z
-      || (c >= 97 && c <= 122) // a-z
-      || (c >= 48 && c <= 57) // 0-9
-      || ch === '_'
-      || ch === '-'
-      || ch === ':'
-    )
-  }
-
-  const trimStartIndentWs = (s: string) => {
-    let i = 0
-    while (i < s.length && isIndentWs(s[i])) i++
-    return s.slice(i)
-  }
-
-  const parseBlockquotePrefix = (rawLine: string) => {
-    let i = 0
-    let saw = false
-    let prefixEnd = 0
-
-    while (i < rawLine.length) {
-      // allow indentation before every marker
-      while (i < rawLine.length && isIndentWs(rawLine[i])) i++
-      if (i >= rawLine.length || rawLine[i] !== '>')
-        break
-      saw = true
-      i++ // consume '>'
-      while (i < rawLine.length && isIndentWs(rawLine[i])) i++
-      prefixEnd = i
-    }
-
-    if (!saw)
-      return null
-
-    const prefix = rawLine.slice(0, prefixEnd)
-    const key = prefix.replace(/[ \t]+$/, '')
-    return {
-      prefix,
-      key,
-      content: rawLine.slice(prefixEnd),
-    }
-  }
-
-  // Keep behavior conservative: only insert a blank line before a custom tag
-  // when it follows a non-blank, non-HTML-ish line. This fixes the common case:
-  //
-  //   paragraph text
-  //   <CustomTag>...</CustomTag>
-  //
-  // Without the blank line, CommonMark HTML block type 7 cannot interrupt a
-  // paragraph, so markdown-it tokenizes the tag as inline HTML inside the
-  // paragraph.
-  const previousLineLooksHtmlish = (line: string) => {
-    const trimmed = trimStartIndentWs(line)
-    return trimmed.startsWith('<')
-  }
-
-  const lineIsBlank = (line: string) => {
-    for (let i = 0; i < line.length; i++) {
-      const ch = line[i]
-      if (ch !== ' ' && ch !== '\t')
-        return false
-    }
-    return true
-  }
-
-  const parseOpeningCustomTagName = (line: string) => {
-    if (isIndentedCodeLine(line))
-      return ''
-    const trimmed = trimStartIndentWs(line)
-    if (!trimmed.startsWith('<'))
-      return ''
-
-    let i = 1
-    while (i < trimmed.length && isIndentWs(trimmed[i])) i++
-    if (i >= trimmed.length)
-      return ''
-    if (trimmed[i] === '/' || trimmed[i] === '!' || trimmed[i] === '?')
-      return ''
-
-    const nameStart = i
-    while (i < trimmed.length && isNameChar(trimmed[i])) i++
-    if (i === nameStart)
-      return ''
-
-    const name = trimmed.slice(nameStart, i).toLowerCase()
-    if (!tagSet.has(name))
-      return ''
-
-    // Require a boundary after tag name to avoid matching prefixes.
-    const next = trimmed[i]
-    if (next && next !== ' ' && next !== '\t' && next !== '>' && next !== '/')
-      return ''
-
-    return name
-  }
-
-  const parseLineStartCustomTag = (line: string) => {
-    if (isIndentedCodeLine(line))
-      return null
-    const trimmed = trimStartIndentWs(line)
-    if (!trimmed.startsWith('<'))
-      return null
-
-    let i = 1
-    while (i < trimmed.length && isIndentWs(trimmed[i])) i++
-    if (i >= trimmed.length)
-      return null
-
-    const isClose = trimmed[i] === '/'
-    if (isClose) {
-      i++
-      while (i < trimmed.length && isIndentWs(trimmed[i])) i++
-    }
-    // Ignore non-element markup (comments/doctypes/pi)
-    const next = trimmed[i]
-    if (!next || next === '!' || next === '?')
-      return null
-
-    const nameStart = i
-    while (i < trimmed.length && isNameChar(trimmed[i])) i++
-    if (i === nameStart)
-      return null
-
-    const name = trimmed.slice(nameStart, i).toLowerCase()
-    if (!tagSet.has(name))
-      return null
-
-    // Require boundary after name so we don't match prefixes
-    const boundary = trimmed[i]
-    if (boundary && boundary !== ' ' && boundary !== '\t' && boundary !== '>' && boundary !== '/')
-      return null
-
-    if (isClose)
-      return { type: 'close' as const, name }
-
-    // opening tag: treat "<tag .../>" as complete on one line
-    if (/\/\s*>\s*$/.test(trimmed))
-      return { type: 'open' as const, name, complete: true as const }
-
-    const gt = trimmed.indexOf('>', i)
-    if (gt !== -1) {
-      const after = trimmed.slice(gt + 1)
-      const closeRe = new RegExp(`<\\s*\\/\\s*${name}\\s*>`, 'i')
-      if (closeRe.test(after))
-        return { type: 'open' as const, name, complete: true as const }
-    }
-
-    return { type: 'open' as const, name, complete: false as const }
-  }
-
-  const parseStandaloneCompleteHtmlTagLine = (line: string) => {
-    if (isIndentedCodeLine(line))
-      return null
-
-    const trimmed = trimStartIndentWs(line).replace(/[ \t]+$/, '')
-    if (!trimmed.startsWith('<'))
-      return null
-    if (/^<\s*(?:!--|!doctype\b|\?)/i.test(trimmed))
-      return null
-
-    const selfClosingMatch = trimmed.match(/^<\s*([A-Z][\w:-]*)\b[^>]*\/\s*>\s*$/i)
-    if (selfClosingMatch?.[1])
-      return selfClosingMatch[1].toLowerCase()
-
-    const fullMatch = trimmed.match(/^<\s*([A-Z][\w:-]*)\b[^>]*>[\s\S]*<\s*\/\s*([A-Z][\w:-]*)\s*>\s*$/i)
-    if (!fullMatch?.[1] || !fullMatch[2])
-      return null
-
-    const openTag = fullMatch[1].toLowerCase()
-    const closeTag = fullMatch[2].toLowerCase()
-    return openTag === closeTag ? openTag : null
-  }
-
-  // Track fenced code blocks so we don't touch their contents.
-  let inFence = false
-  let fenceChar: '`' | '~' | '' = ''
-  let fenceLen = 0
-
-  const parseFenceMarker = (line: string) => {
-    let i = 0
-    while (i < line.length && isIndentWs(line[i])) i++
-    const ch = line[i]
-    if (ch !== '`' && ch !== '~')
-      return null
-    let j = i
-    while (j < line.length && line[j] === ch) j++
-    const len = j - i
-    if (len < 3)
-      return null
-    return { markerChar: ch as '`' | '~', markerLen: len, rest: line.slice(j) }
-  }
-
-  const fenceMatchLine = (rawLine: string) => parseFenceMarker(rawLine)
-
-  const lineStartsWithBlockMarker = (line: string) => {
-    const trimmed = trimStartIndentWs(line)
-    if (!trimmed)
-      return false
-    if (isIndentedCodeLine(line))
-      return true
-    return /^(?:#{1,6}[ \t]+|>|[*+-][ \t]+|\d+[.)][ \t]+|`{3,}|~{3,}|\||\$\$|:{3,}|\[\^[^\]]+\]:|-{3,}|\*{3,}|_{3,})/.test(trimmed)
-  }
-
-  const currentCustomBlockNeedsBoundary = (lineStart: number, currentQuoteKey: string, tagName: string) => {
-    let scanIdx = lineStart
-    let depth = 0
-
-    while (scanIdx < markdown.length) {
-      const nl = markdown.indexOf('\n', scanIdx)
-      const hasNl = nl !== -1
-      const isCrlf = hasNl && nl > scanIdx && markdown[nl - 1] === '\r'
-      const lineEnd = hasNl ? (isCrlf ? nl - 1 : nl) : markdown.length
-      const rawLine = markdown.slice(scanIdx, lineEnd)
-
-      const blockquote = parseBlockquotePrefix(rawLine)
-      const quoteKey = blockquote?.key ?? ''
-      if (depth > 0 && currentQuoteKey && quoteKey !== currentQuoteKey)
-        break
-
-      const contentLine = blockquote?.content ?? rawLine
-      const lineTag = parseLineStartCustomTag(contentLine)
-
-      if (lineTag?.name === tagName) {
-        if (lineTag.type === 'open') {
-          if (!lineTag.complete)
-            depth++
-        }
-        else if (depth > 0) {
-          depth--
-          if (depth === 0)
-            return false
-        }
-      }
-      else if (depth > 0) {
-        if (lineIsBlank(contentLine) || lineStartsWithBlockMarker(contentLine))
-          return true
-      }
-
-      if (hasNl)
-        scanIdx = nl + 1
-      else
-        break
-    }
-
-    return false
-  }
-
-  let out = ''
-  let idx = 0
-  let prevLineBlank = true
-  let prevLineHtmlish = false
-  let prevLineStandaloneCompleteHtmlTag = false
-  // Use the last seen newline sequence to insert a blank line that matches the file.
-  let lastNewline = '\n'
-  const customBlockStack: string[] = []
-  let prevQuoteKey = ''
-
-  while (idx < markdown.length) {
-    const nl = markdown.indexOf('\n', idx)
-    const hasNl = nl !== -1
-    const isCrlf = hasNl && nl > idx && markdown[nl - 1] === '\r'
-    const lineEnd = hasNl ? (isCrlf ? nl - 1 : nl) : markdown.length
-    const line = markdown.slice(idx, lineEnd)
-    const newline = hasNl ? (isCrlf ? '\r\n' : '\n') : ''
-
-    const blockquote = parseBlockquotePrefix(line)
-    const quoteKey = blockquote?.key ?? ''
-    const contentLine = blockquote?.content ?? line
-
-    // Maintain fence state based on the original line.
-    const fenceMatch = fenceMatchLine(contentLine)
-    if (fenceMatch) {
-      if (inFence) {
-        if (fenceMatch.markerChar === fenceChar && fenceMatch.markerLen >= fenceLen) {
-          if (/^\s*$/.test(fenceMatch.rest)) {
-            inFence = false
-            fenceChar = ''
-            fenceLen = 0
-          }
-        }
-      }
-      else {
-        inFence = true
-        fenceChar = fenceMatch.markerChar
-        fenceLen = fenceMatch.markerLen
-      }
-    }
-
-    const insideCustomBlock = customBlockStack.length > 0
-    if (!inFence && !insideCustomBlock) {
-      const opening = parseOpeningCustomTagName(contentLine)
-      const needsBoundaryAfterStandaloneHtml
-        = !!opening
-          && !prevLineBlank
-          && prevLineHtmlish
-          && prevLineStandaloneCompleteHtmlTag
-          && currentCustomBlockNeedsBoundary(idx, quoteKey, opening)
-      if (opening && !prevLineBlank && (!prevLineHtmlish || needsBoundaryAfterStandaloneHtml)) {
-        // Insert a blank line boundary between the previous paragraph line and the custom block.
-        // In blockquotes, the blank line must also carry the `>` markers, otherwise the
-        // blockquote would end and the tag would escape the quote.
-        if (quoteKey && prevQuoteKey && quoteKey === prevQuoteKey) {
-          out += `${quoteKey}${lastNewline}`
-        }
-        else if (!quoteKey) {
-          out += lastNewline
-        }
-      }
-    }
-
-    out += line
-    out += newline
-
-    if (newline)
-      lastNewline = newline
-
-    // Maintain custom-tag "block stack" only when not inside fenced code.
-    // This avoids accidentally inserting blank lines inside <CustomTag> blocks
-    // which would mutate their captured inner content.
-    if (!inFence) {
-      const tag = parseLineStartCustomTag(contentLine)
-      if (tag) {
-        if (tag.type === 'open') {
-          if (!tag.complete)
-            customBlockStack.push(tag.name)
-        }
-        else {
-          // Close: pop matching tag (or unwind to it if nesting is malformed)
-          for (let j = customBlockStack.length - 1; j >= 0; j--) {
-            if (customBlockStack[j] === tag.name) {
-              customBlockStack.length = j
-              break
-            }
-          }
-        }
-      }
-    }
-
-    // Update "previous line" info for the next iteration (based on the original line).
-    const blank = lineIsBlank(contentLine)
-    prevLineBlank = blank
-    prevLineHtmlish = !blank && previousLineLooksHtmlish(contentLine)
-    prevLineStandaloneCompleteHtmlTag = !blank && !!parseStandaloneCompleteHtmlTagLine(contentLine)
-    prevQuoteKey = quoteKey
-
-    idx = hasNl ? nl + 1 : markdown.length
-  }
-
-  return out
-}
-
-/**
- * Math-context scanner used to gate the transport-split LaTeX-command
- * reconstruction. Tracks fenced code (including list/blockquote-prefixed
- * fences via matchMarkdownFenceMarker), backtick spans, $$...$$ blocks,
- * \[...\] spans and single-$ math.
- *
- * Single-$ math is line-scoped (markdown-it-math never lets $...$ span a
- * line break): the opener state resets at every line break. A $ that is the
- * last char of a line counts as an opener because a transport split can land
- * exactly after it (`$\nabla$` -> "$" + newline + "abla$").
- *
- * The scanner is a pure function of the consumed text, so the append-only
- * windowed fast path in getSafeMarkdown stays deterministic.
- */
-function createLatexSplitMathScanner(source: string) {
-  let inFence = false
-  let fenceMarker: '`' | '~' | '' = ''
-  let fenceLen = 0
-  let inDollarBlock = false
-  let inBracketMath = false
-  let singleDollarOpen = false
-  let scanned = 0
-
-  const processLine = (line: string, _lineStartOffset: number) => {
-    const fenceMatch = matchMarkdownFenceMarker(line)
-    if (fenceMatch) {
-      if (inFence) {
-        if (fenceMatch.markerChar === fenceMarker && fenceMatch.markerLen >= fenceLen && /^\s*$/.test(fenceMatch.rest)) {
-          inFence = false
-          fenceMarker = ''
-          fenceLen = 0
-        }
-      }
-      else {
-        inFence = true
-        fenceMarker = fenceMatch.markerChar
-        fenceLen = fenceMatch.markerLen
-      }
-      return
-    }
-
-    if (inFence)
-      return
-
-    let i = 0
-    while (i < line.length) {
-      if (inDollarBlock) {
-        if (line.startsWith('$$', i) && !isEscapedDelimiterAt(line, i)) {
-          inDollarBlock = false
-          i += 2
-        }
-        else {
-          i++
-        }
-        continue
-      }
-
-      if (inBracketMath) {
-        if (line.startsWith('\\]', i) && !isEscapedDelimiterAt(line, i)) {
-          inBracketMath = false
-          i += 2
-        }
-        else {
-          i++
-        }
-        continue
-      }
-
-      const ch = line[i]
-      if (ch === '`') {
-        const runLen = countRepeatedChar(line, i, '`')
-        const closeIndex = findCodeSpanCloseIndex(line, i + runLen, runLen)
-        if (closeIndex === -1)
-          break // unclosed span: the rest of the line is code
-        i = closeIndex + runLen
-        continue
-      }
-
-      if (ch === '\\') {
-        const next = line[i + 1]
-        if (next === '[' && !isEscapedDelimiterAt(line, i)) {
-          inBracketMath = true
-          i += 2
-        }
-        else if (next === ']' && !isEscapedDelimiterAt(line, i) && !inBracketMath) {
-          i += 2
-        }
-        else {
-          i += 2 // escaped char (\$, \`, \\, ...)
-        }
-        continue
-      }
-
-      if (ch === '$') {
-        if (line[i + 1] === '$' && !isEscapedDelimiterAt(line, i)) {
-          inDollarBlock = true
-          singleDollarOpen = false
-          i += 2
-          continue
-        }
-        if (singleDollarOpen) {
-          singleDollarOpen = false
-          i++
-          continue
-        }
-        // Opener when followed by non-whitespace, or at end of line (a split
-        // can land right after the opener). A $ followed by a digit is
-        // currency in prose ("$5 total") and must not open math.
-        const after = line[i + 1]
-        if (after === undefined || ((after !== ' ' && after !== '\t') && !/\d/.test(after)))
-          singleDollarOpen = true
-        i++
-        continue
-      }
-
-      i++
-    }
-  }
-
-  const scanTo = (target: number) => {
-    while (scanned < target) {
-      const newlineIndex = source.indexOf('\n', scanned)
-      const lineEndRaw = newlineIndex === -1 || newlineIndex >= target ? target : newlineIndex
-      const lineEnd = lineEndRaw > scanned && source[lineEndRaw - 1] === '\r' ? lineEndRaw - 1 : lineEndRaw
-      const line = source.slice(scanned, lineEnd)
-      processLine(line, scanned)
-      if (newlineIndex === -1 || newlineIndex >= target) {
-        scanned = target
-        break
-      }
-      singleDollarOpen = false // single-$ math never spans a line
-      scanned = newlineIndex + 1
-    }
-  }
-
-  return {
-    scanTo,
-    inMath: () => inDollarBlock || inBracketMath || singleDollarOpen,
-  }
-}
-
-function transformStreamingSafeMarkdown(
-  source: string,
-  isFinal: boolean,
-  md: MarkdownIt,
-  options: ParseOptions,
-) {
-  // Reconstruct transport-split LaTeX commands ONLY inside open math contexts
-  // ($...$ / $$...$$ / \[...\]). The previous version rewrote every soft line
-  // break followed by `abla|eq|ot|exists`, which corrupted ordinary prose
-  // ("First.\nother things" gained a literal backslash-n and merged
-  // paragraphs). A bare CR (the \rho/\right split) cannot occur in prose, so
-  // that reconstruction stays ungated.
-  let safeMarkdown = source.replace(/([^\\])\r(ight|ho)/g, '$1\\r$2')
-  const latexSplitMathScanner = createLatexSplitMathScanner(safeMarkdown)
-  safeMarkdown = safeMarkdown.replace(/([^\\])\r?\n(abla|eq|ot|exists)/g, (full, before, cmd, offset) => {
-    // Consume up to (not including) the newline at offset+1: the $ that opens
-    // math may sit anywhere on this line, including as its last char.
-    latexSplitMathScanner.scanTo(offset + 1)
-    const shouldReconstruct = latexSplitMathScanner.inMath()
-    // Consume the newline itself: single-$ math is line-scoped and resets.
-    latexSplitMathScanner.scanTo(offset + full.length)
-    return shouldReconstruct ? `${before}\\n${cmd}` : full
-  })
-
-  if (!isFinal) {
-    if (safeMarkdown.endsWith('- *')) {
-      // 放置markdown 解析 - * 会被处理成多个 ul >li 嵌套列表
-      safeMarkdown = safeMarkdown.replace(/- \*$/, '- \\*')
-    }
-    if (/(?:^|\n)\s*-\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：单独的 "-" 行（或以换行结尾的 "-\n"）会被渲染成文本/列表前缀，
-      // 也会导致输入 "---" 时第一个 "-" 先闪出来再跳成 hr。
-      safeMarkdown = safeMarkdown.replace(/(?:^|\n)\s*-\s*$/, (m) => {
-        return m.startsWith('\n') ? '\n' : ''
-      })
-    }
-    else if (/(?:^|\n)\s*--\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：输入 "---" 时的 "--" 前缀也不应该作为文本渲染，避免跳动。
-      safeMarkdown = safeMarkdown.replace(/(?:^|\n)\s*--\s*$/, (m) => {
-        return m.startsWith('\n') ? '\n' : ''
-      })
-    }
-    else if (/(?:^|\n)\s*>\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：单独的 ">" 行会先被识别成 blockquote，导致 UI 闪烁/跳动。
-      // 只裁剪末尾这一个 marker，等后续内容到齐再正常解析。
-      safeMarkdown = safeMarkdown.replace(/(?:^|\n)\s*>\s*$/, (m) => {
-        return m.startsWith('\n') ? '\n' : ''
-      })
-    }
-    else if (/\n\s*[*+]\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：单独的 "*"/"+" 行会被识别成空的 list item，导致 UI 闪出一个圆点
-      safeMarkdown = safeMarkdown.replace(/\n\s*[*+]\s*$/, '\n')
-    }
-    else if (/(?:^|\n)\s*\d+\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：单独的 "2" / "10" 行常是有序列表 marker 的前缀（下一字符才到 "." / ")"）。
-      // 在此状态下 markdown-it 会把它解析成 paragraph/text，导致先撑开一段空白再被下一次解析替换，形成抖动。
-      // 只裁剪末尾这一行，等 marker 完整或有内容后再正常解析。
-      // 但当整个文档本身就是纯数字（例如 "1234567"）时，这不是列表前缀，而是正常文本内容，
-      // 不应被裁剪为空，否则会导致 parse 结果一直为空。
-      if (!/^\d+$/.test(safeMarkdown.trim())) {
-        safeMarkdown = safeMarkdown.replace(/(?:^|\n)\s*\d+\s*$/, (m) => {
-          return m.startsWith('\n') ? '\n' : ''
-        })
-      }
-    }
-    else if (/(?:^|\n)\s*\d+[.)]\s+\*{1,3}\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：有序列表项刚开始输出 "**"（粗体）时，常会经历 "1. *" / "1. **" 等尾部状态。
-      // markdown-it 在这些状态下可能把 "*" 当作空的 bullet list marker（嵌套列表），导致 UI 先闪一个圆点/空块再恢复。
-      // 将尾部孤立的星号临时转义，避免被当作列表 marker。
-      safeMarkdown = safeMarkdown.replace(
-        /((?:^|\n)\s*\d+[.)]\s+)(\*{1,3})\s*$/,
-        (_, prefix: string, stars: string) => `${prefix}${stars.split('').map(() => '\\*').join('')}`,
-      )
-    }
-    else if (/(?:^|\n)\s*\d+[.)]\s*$/.test(safeMarkdown)) {
-      // streaming 中间态：单独的 "2." / "3)" 行会先被渲染成列表/段落占位，随后合并成真正的 list item，导致抖动。
-      // 裁剪末尾这一个 marker，等后续内容到齐再正常解析。
-      safeMarkdown = safeMarkdown.replace(/(?:^|\n)\s*\d+[.)]\s*$/, (m) => {
-        return m.startsWith('\n') ? '\n' : ''
-      })
-    }
-    else if (/\n[[(]\n*$/.test(safeMarkdown)) {
-      // 此时 markdown 解析会出错要跳过
-      safeMarkdown = safeMarkdown.replace(/(\n\[|\n\()+\n*$/g, '\n')
-    }
-
-    // The tolerant-math boundary scan is applied by the caller to the full
-    // document (it carries incremental fence/math state across commits).
-    safeMarkdown = getStreamingAdmonitionOpenTailReplacement(safeMarkdown, options.customHtmlTags) ?? safeMarkdown
-  }
-
-  // For custom HTML-like blocks (e.g. <thinking>...</thinking>), markdown-it may
-  // keep parsing subsequent lines as part of the HTML block unless there's a
-  // blank line boundary. To ensure content immediately following a closing tag
-  // (like a list/table/blockquote/fence) is parsed as Markdown blocks, insert
-  // a single empty line after the closing tag when the next line begins with a
-  // block-level marker.
-  if (options.customHtmlTags?.length && safeMarkdown.includes('<')) {
-    const tags = normalizeCustomHtmlTags(options.customHtmlTags)
-
-    if (tags.length) {
-      safeMarkdown = ensureBlankLineBeforeInlineMultilineCustomHtmlBlocks(safeMarkdown, tags)
-      // markdown-it doesn't always treat custom tags as html_block when the opening
-      // tag and the first content token live on the same line (e.g. "<thinking> foo").
-      // That causes the tag to be parsed as inline HTML and breaks custom block parsing.
-      // Normalize "<tag> ..." (line-start only) into "<tag>\n..." so it becomes a block.
-      safeMarkdown = normalizeCustomHtmlOpeningTagSameLine(safeMarkdown, tags)
-      // CommonMark HTML blocks of type 7 cannot interrupt paragraphs. When a custom
-      // tag line (e.g. "<RadioBtn>") immediately follows paragraph text, markdown-it
-      // will tokenize it as inline HTML and merge it into the paragraph. Insert a
-      // blank line boundary before custom tags that follow non-HTML-ish text lines.
-      safeMarkdown = ensureBlankLineBeforeCustomHtmlBlocks(safeMarkdown, tags)
-      // In streaming output, models sometimes emit "</tag>## Heading" without a
-      // newline after the custom block close. Split it into separate lines so the
-      // "##" can be parsed as a heading (and to avoid being swallowed by HTML block parsing).
-      safeMarkdown = ensureBlankLineAfterCustomHtmlCloseBeforeBlockMarkerSameLine(safeMarkdown, tags)
-
-      // Fast path: no closing tag marker at all.
-      if (!safeMarkdown.includes('</')) {
-        // no-op
-      }
-      else {
-        for (const tag of tags) {
-          const re = new RegExp(
-          // After a closing tag at end-of-line, if the next line is not blank
-          // (ignoring whitespace) and we're not at end-of-string, insert a
-          // blank line to force markdown-it to resume normal block parsing.
-          // Restrict to lines that contain ONLY the closing tag (plus whitespace)
-          // to avoid affecting inline occurrences like "x</thinking>y".
-            String.raw`(^[\t ]*<\s*\/\s*${tag}\s*>[\t ]*)(\r?\n)(?![\t ]*\r?\n|$)`,
-            'gim',
-          )
-          safeMarkdown = safeMarkdown.replace(re, '$1$2$2')
-        }
-      }
-    }
-  }
-
-  // 마지막에 남아있는 미완성 '<...'(예: '<fo', '</think') 꼬리 조각은
-  // streaming 중간 상태에서 화면에 그대로 찍힐 수 있으므로, markdown-it
-  // 파싱 전에 제거한다.
-  if (!isFinal)
-    safeMarkdown = stripDanglingHtmlLikeTail(safeMarkdown)
-
-  return safeMarkdown
-}
-
-function getSafeMarkdown(md: MarkdownIt, sourceMarkdown: string, isFinal: boolean, options: ParseOptions) {
-  const owner = md as unknown as object
-  const mode = `${isFinal ? 'final' : 'stream'}:${(options.customHtmlTags ?? []).join(',')}`
-  const previous = safeMarkdownCache.get(owner)
-
-  let safeMarkdown: string
-  if (
-    // Append-only streaming fast path: only the appended tail can introduce
-    // new mid-state markers, and the prefix of the previous safe markdown was
-    // already transformed identically. Transform a tail window of the RAW
-    // source (old tail margin + appended chunk) and stitch it onto the cached
-    // prefix, avoiding O(doc) regex scans on every commit (quadratic total
-    // for long streaming documents).
-    !isFinal
-    && !options.customHtmlTags?.length
-    && previous
-    && previous.mode === mode
-    && sourceMarkdown.length >= previous.source.length
-    && sourceMarkdown.startsWith(previous.source)
-  ) {
-    // The window cut MUST be a raw-source index. The previous implementation
-    // cut `previous.safeMarkdown` at a safeMarkdown index but sliced the raw
-    // source with it: any char-inserting fix earlier in the document (e.g.
-    // `\n(abla|eq|ot|exists)` / `\r(ight|ho)`) made the two index spaces
-    // diverge and silently dropped/repeated chars at the seam.
-    const windowStart = Math.max(0, previous.source.length - SAFE_MARKDOWN_WINDOW_MARGIN - SAFE_MARKDOWN_WINDOW_OVERLAP)
-    const window = sourceMarkdown.slice(windowStart)
-    const transformed = transformStreamingSafeMarkdown(window, isFinal, md, options)
-    // Overlap verification: the part of the window that overlaps the previous
-    // source (its first `previous.source.length - windowStart` chars) must be
-    // byte-identical to the previous safe markdown tail. A re-transform of
-    // unchanged source is deterministic, so equality proves the overlap region
-    // had no char-inserting fixes (which would shift the seam) AND that the
-    // appended chunk did not complete a cross-boundary fix (e.g. a trailing
-    // `\r` followed by an appended `ight`). Any divergence falls back to a
-    // full-document transform, which is always correct.
-    const overlapLength = previous.source.length - windowStart
-    const overlapOk = transformed.length >= overlapLength
-      && transformed.slice(0, overlapLength) === previous.safeMarkdown.slice(-overlapLength)
-    safeMarkdown = overlapOk
-      ? previous.safeMarkdown.slice(0, previous.safeMarkdown.length - overlapLength) + transformed
-      : transformStreamingSafeMarkdown(sourceMarkdown, isFinal, md, options)
-  }
-  else {
-    safeMarkdown = transformStreamingSafeMarkdown(sourceMarkdown, isFinal, md, options)
-  }
-
-  // The tolerant-math boundary scan carries incremental fence/math state in
-  // its own cache keyed by md, so it must always run on the full document
-  // (not a tail window) — otherwise pre-window math openers would be
-  // invisible and ambiguous tails would not be hidden.
-  if (!isFinal)
-    safeMarkdown = stripPendingExplicitMathTail(safeMarkdown, md)
-
-  safeMarkdownCache.set(owner, { source: sourceMarkdown, safeMarkdown, mode })
-  return safeMarkdown
-}
-
 export function parseMarkdownToStructure(
   markdown: string,
   md: MarkdownIt,
@@ -3830,7 +1614,7 @@ export function parseMarkdownToStructure(
     // The safe-markdown cache is owned by the top-level streaming session;
     // a final auto-parse ends that session, so drop the retained source +
     // transform (the next stream parse starts a fresh document).
-    safeMarkdownCache.delete(md as unknown as object)
+    clearSafeMarkdownCache(md)
   }
 
   const safeMarkdown = getSafeMarkdown(md, sourceMarkdown, isFinal, options)
